Keep column arrow up to date
When an order is set on a One2Many widget and the user reload the form, the
selected order is lost and so the client should no more display those arrow.
For that we update the arrows on each display using the screen order. And we
unset the screen order when the group change as we do not know the order of
this group.

issue7817
review70481002