Remove unnecessary calls to fields_view_get
Screen.count_tab_domain should be called after the switch_view as it is done
in tryton. This avoid to trigger a call to fields_view_get for the
domain_parser.
Screen.switch_view is performing one extra loop compared to tryton. This
trigger and extra call fields_view_get.

issue8167
review60581002