Ensure to put search_value in URL
If the screen container or the domain parser are not yet ready to compute the
search_value, we should ensure to use the initial search_value. Otherwise sao
may trigger to open a new tab without the search_value because the URL will be
set with an empty search.

issue8267
review281471002